Moved Step1 to deprecated. Change File and folder name for all files
